Add support for optional SQS arguments
FIFO queues are required to have a MessageGroupID, and a DeduplicationID if the queue is not set to have content-based deduplication. As we cannot provide our own arguments to SQS through the library, this makes it impossible to write to FIFO queues. This PR allows one to submit a dictionary of key-value pairs which will be sent as arguments, directly to the SQS library.